tcp_metrics: use a flex_array for the tcp_metrics hash table
On a system with moderate amounts of memory (> 128*1024*4K) the
tcpm_hash_bucket array is sized with 16*1024*sizeof(void*) by default.

This results in a 128K kmalloc allocation, which on a system with 4K
pages translates into an order-5 get_free_pages. This is all well and
good most of the time, but on a system with NETNS enabled, a new hash
table is allocated every time we clone(CLONE_NEWNS|...) and we quickly
exhaust available order-5 slabs. Chrome's sandbox for instance leads to
exhaustion rather quickly.

Instead of using vmalloc which may exhaust vmalloc space on i386 kernels,
and can result in a lot of slow page tables muckery, use <linux/flex_array.h>
which breaks up the contiguous region into 32 order-0 allocations (with 4K
pages.) that are much easier to satisfy as fragmentation increases.

I toyed with the ideas of:
 - using vmalloc
 - making the hash_table global instead of pernet and tagging entries with
   the net-pointer to avoid cross-netns pollution and the privacy implications
   of figuring out which name spaces have been talking to which addrs.
 - reducing the number of buckets and increasing the chain length
and using flex_array seemed like the best option.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>