[HACKS] Allow building SRPM on OL8 outside of Oracle's environment
Building the SRPM on Fedora 42 redefines %rpmversion... so give it
the same name as in the Fedora kernel.spec. Similarly also disable some
sub-packages which are unnecessary as well as turning off kabichk.
Finally, the redhat-rpm-config package contains the macros prior to the
existence of kernel-{rpm,srpm}-macros packages, so remove that
dependency, and an updated dwarves does not seem to be necessary either.

With all this, we're successfully building binary RPMs for the kernel
using the oraclelinux-8-x86_64.cfg mock config.

This is all necessary to support Hammerspace testing on OL8.

Signed-off-by: Kyle McMartin <jkkm@jkkm.org>